feat(refactor): refactor to allow generic usage of komokana
